requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
langchain==0.2.6
langchain-community==0.2.6
pypdf==4.2.0
tiktoken==0.7.0
rich==13.7.1
click==8.1.7
watchdog==4.0.1
streamlit==1.36.0
ollama==0.2.1
python-dotenv==1.0.1
numpy==1.26.4
